\name{basePlot-methods}
\docType{methods}
\alias{basePlot-methods}
\alias{basePlot,list-method}
\alias{basePlot,SequenceSummary-method}
\title{}
\description{
  
}

\usage{

}

\arguments{
  \item{obj}{an S4 object of class \code{FASTQSummary} from \code{readSeqFile}.}
}

\section{Methods}{
\describe{

\item{\code{signature(x = "list")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "SequenceSummary")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
